feat: refactor RecentBoards to Boards component to reuseable component
