+++ src/components/AppBar/AppBar.scss
@@ -127,147 +127,4 @@ $app-bar-max-height: 2rem;
   }
 }
 
-.workspace {
-  padding: 0.75rem 0;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-  // padding-bottom: 0.75rem;
-  .workspace__title {
-    color: $text-sub-title-color;
-    font-size: 0.9rem;
-    margin-bottom: 0.75rem;
-  }
-  .workspace__details {
-    display: flex;
-    gap: 0.75rem;
-    align-items: center;
-    border-radius: 3px;
 
-    &:hover {
-      cursor: pointer;
-      background-color: rgba(0, 0, 0, 0.1);
-    }
-
-    &.workspace__details--current {
-      &:hover {
-        cursor: auto;
-        background-color: inherit;
-      }
-    }
-
-    .workspace__avatar {
-      flex: 0 0 auto;
-      width: 2.6rem;
-      height: 2.6rem;
-      background: linear-gradient(#006644, #00875a);
-      text-align: center;
-      border-radius: inherit;
-      span {
-        line-height: 2.6rem;
-        font-size: 2rem;
-        font-weight: 700;
-        color: #ffffff;
-      }
-    }
-    .workspace__name {
-      flex-grow: 1;
-      color: $app-main-text-color;
-      font-weight: 700;
-      font-size: 1rem;
-    }
-  }
-}
-
-.recent-board {
-  padding: 0.75rem 0;
-  > * {
-    margin-bottom: 0.5rem;
-  }
-
-  .board {
-    display: flex;
-    align-items: center;
-    height: 2rem;
-
-    font-weight: 700;
-    border-radius: 3px;
-
-    &:hover {
-      cursor: pointer;
-      background-color: rgba(0, 0, 0, 0.1);
-    }
-
-    &:hover > .board-thumb {
-      opacity: 0.6;
-    }
-
-    &:hover > .board-star {
-      display: inline;
-    }
-
-    .board-thumb {
-      flex-shrink: 0;
-      width: 40px;
-      height: inherit;
-      background-size: cover;
-      border-radius: inherit;
-    }
-
-    .board-info {
-      padding-left: 0.5rem;
-      min-width: 100px;
-      flex: 1;
-      div {
-        font-size: 14px;
-        color: $app-main-text-color;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-      }
-
-      span {
-        font-size: 12px;
-        font-weight: 100;
-        padding-right: 0.5rem;
-        color: $text-sub-title-color;
-      }
-    }
-
-    .board-template-btn {
-      height: 1.375rem;
-      padding: 3px 0.5rem;
-      margin: 0 0.5rem 0 0;
-      font-size: 11px;
-      font-weight: inherit;
-    }
-
-    .board-star {
-      display: none;
-      border: none;
-      outline: none;
-      margin-right: 0.375rem;
-      width: 1.2rem;
-      height: 1.2rem;
-
-      &:hover {
-        color: #f2d600;
-      }
-
-      &.board-star--starred {
-        display: block;
-        color: #f2d600;
-      }
-    }
-  }
-}
-
-ul.recent-board {
-  list-style: none;
-  margin: 0;
-
-  li {
-    margin-bottom: 0.5rem;
-    &:last-child {
-      margin: 0;
-    }
-  }
-}
