+++ src/components/Workspaces/Workspaces.scss
@@ -1,0 +1,51 @@
+@import "_setting.scss";
+
+.workspace {
+  padding: 0.75rem 0;
+  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+  // padding-bottom: 0.75rem;
+  .workspace__title {
+    color: $text-sub-title-color;
+    font-size: 0.9rem;
+    margin-bottom: 0.75rem;
+  }
+  .workspace__details {
+    display: flex;
+    gap: 0.75rem;
+    align-items: center;
+    border-radius: 3px;
+
+    &:hover {
+      cursor: pointer;
+      background-color: rgba(0, 0, 0, 0.1);
+    }
+
+    &.workspace__details--current {
+      &:hover {
+        cursor: auto;
+        background-color: inherit;
+      }
+    }
+
+    .workspace__avatar {
+      flex: 0 0 auto;
+      width: 2.6rem;
+      height: 2.6rem;
+      background: linear-gradient(#006644, #00875a);
+      text-align: center;
+      border-radius: inherit;
+      span {
+        line-height: 2.6rem;
+        font-size: 2rem;
+        font-weight: 700;
+        color: #ffffff;
+      }
+    }
+    .workspace__name {
+      flex-grow: 1;
+      color: $app-main-text-color;
+      font-weight: 700;
+      font-size: 1rem;
+    }
+  }
+}
